Creados Aliados y Enemigos Normales y cambio a uint16_t de los id

+++ src/src/IA/Decision_Manager.cpp
@@ -19,11 +19,10 @@ Decision_Manager::Decision_Manager(Action_Manager* _i_action_manager) {
 
 	_enemigos = _interfaz_decision->get_enemigos();
 
-	unsigned short _n_enemigos = _interfaz_decision->get_n_enemigos();
-	//unsigned short _n_enemigos = 0;
+	uint16_t _n_enemigos = _interfaz_decision->get_n_enemigos();
 	_blackboards = new Blackboard*[_n_enemigos];
 
-	for(uint8_t _cont = 0; _cont<_n_enemigos; _cont++) {
+	for(uint16_t _cont = 0; _cont<_n_enemigos; _cont++) {
 		_blackboards[_cont] = new Blackboard(*_interfaz_decision, _cont);
 		_enemigos[_cont]->set_blackboard(_blackboards[_cont]);
 	}
@@ -50,8 +49,8 @@ Decision_Manager::Decision_Manager(Action_Manager* _i_action_manager) {
 Decision_Manager::~Decision_Manager() {
 	delete _decision_tree_manager;
 
-	unsigned short _n_enemigos = _interfaz_decision->get_n_enemigos();
-	for(uint8_t _cont=0; _cont<_n_enemigos; _cont++) {
+	uint16_t _n_enemigos = _interfaz_decision->get_n_enemigos();
+	for(uint16_t _cont=0; _cont<_n_enemigos; _cont++) {
 		delete _blackboards[_cont];
 	}
 	delete [] _blackboards;
@@ -64,8 +63,7 @@ Decision_Manager::~Decision_Manager() {
 
 // Recorre todos los enemigos y les agrega una accion
 void Decision_Manager::toma_decisiones(){
-	unsigned short _n_enemigos = _interfaz_decision->get_n_enemigos();
-	//unsigned short _n_enemigos = 0;	
+	uint16_t _n_enemigos = _interfaz_decision->get_n_enemigos();
 	// Control de las iteraciones del level of detail
 	for(uint8_t _cont = 1; _cont<_niveles_lod; _cont++) {
 		_iteraciones_actuales_lod[_cont]++;
@@ -76,7 +74,7 @@ void Decision_Manager::toma_decisiones(){
 	}
 	//std::cout << "\n";
 
-	for(unsigned short _cont=0; _cont<_n_enemigos; _cont++) {
+	for(uint16_t _cont=0; _cont<_n_enemigos; _cont++) {
 
 		// Se actualizan los datos del blackboard
 		_blackboards[_cont]->actualiza_datos();
